Move useMetabotName hook from metabot to common/hooks
useMetabotName is a one-line wrapper around useSetting used by many
modules outside metabot (rich_text_editing, common components,
transforms, enterprise SDK), so it belongs in common/hooks. Moved the
file, exported it from the common/hooks barrel, removed it from the
metabot hooks barrel, and updated every importer, splitting mixed
barrel imports where needed.
