+++ frontend/src/metabase/common/hooks/use-metabot-name.ts
@@ -1,4 +1,4 @@
-import { useSetting } from "metabase/common/hooks";
+import { useSetting } from "./use-setting";
 
 /** Returns the user-configured display name for Metabot (defaults to "Metabot"). */
 export const useMetabotName = (): string => {
