+++ frontend/src/metabase/metabot/components/AIQuestionAnalysisButton.tsx
@@ -1,10 +1,10 @@
 import { t } from "ttag";
 
 import { ToolbarButton } from "metabase/common/components/ToolbarButton";
+import { useMetabotName } from "metabase/common/hooks";
 import { useToast } from "metabase/common/hooks/use-toast";
 import {
   useMetabotAgent,
-  useMetabotName,
   useUserMetabotPermissions,
 } from "metabase/metabot/hooks";
