+++ frontend/src/metabase/metabot/components/MetabotAppBarButton.tsx
@@ -1,9 +1,9 @@
 import { t } from "ttag";
 
+import { useMetabotName } from "metabase/common/hooks";
 import {
   useIsAskPage,
   useMetabotAgent,
-  useMetabotName,
   useUserMetabotPermissions,
 } from "metabase/metabot/hooks";
 import { ActionIcon, type ActionIconProps, Tooltip } from "metabase/ui";
