+++ frontend/src/metabase/metabot/hooks/index.ts
@@ -3,6 +3,5 @@ export { useMetabotEnabledEmbeddingAware } from "./use-metabot-embedding-aware-e
 export { useBranchableMessages } from "./use-branchable-messages";
 export { useMetabotAgent } from "./use-metabot-agent";
 export { useMetabotAgentsManager } from "./use-metabot-agents-manager";
-export { useMetabotName } from "./use-metabot-name";
 export { useDashCardAnalysis } from "./useDashCardAnalysis";
 export { useUserMetabotPermissions } from "./use-user-metabot-permissions";
