+++ frontend/src/metabase/rich_text_editing/tiptap/extensions/MetabotEmbed/MetabotEmbed.tsx
@@ -12,12 +12,10 @@ import { useLatest } from "react-use";
 import { t } from "ttag";
 
 import { useLazyMetabotGenerateContentQuery } from "metabase/api";
+import { useMetabotName } from "metabase/common/hooks";
 import CS from "metabase/css/core/index.css";
 import { MetabotIcon } from "metabase/metabot/components/MetabotIcon";
-import {
-  useMetabotName,
-  useUserMetabotPermissions,
-} from "metabase/metabot/hooks";
+import { useUserMetabotPermissions } from "metabase/metabot/hooks";
 import { useDispatch, useSelector } from "metabase/redux";
 import { useEditorHost } from "metabase/rich_text_editing/tiptap/EditorHost";
 import { Box, Button, Flex, Icon, Text, Tooltip } from "metabase/ui";
