+++ frontend/src/metabase/rich_text_editing/tiptap/extensions/MetabotMention/MetabotSuggestion.tsx
@@ -15,7 +15,7 @@ import type {
   Section,
 } from "metabase/common/components/AccordionList/types";
 import { searchFilter } from "metabase/common/components/AccordionList/utils";
-import { useMetabotName } from "metabase/metabot/hooks";
+import { useMetabotName } from "metabase/common/hooks";
 import type { MenuItem } from "metabase/rich_text_editing/tiptap/extensions/shared/MenuComponents";
 import { MenuItemComponent } from "metabase/rich_text_editing/tiptap/extensions/shared/MenuComponents";
 import { SuggestionPaper } from "metabase/rich_text_editing/tiptap/extensions/shared/SuggestionPaper";
